+++ docs/scripts/api-render.py
@@ -18,6 +18,7 @@ else:
 pdoc.render.configure(
     template_directory=here / "pdoc-template",
     edit_url_map=edit_url_map,
+    search=False,
 )
 # We can't configure Hugo, but we can configure pdoc.
 pdoc.render_helpers.formatter.cssclass = "chroma pdoc-code"
@@ -68,4 +69,14 @@ for module in modules:
         ).encode()
     )
 
-(here / ".." / "src" / "content" / "addons-api.md").touch()
+(here / "../src/content/api/_index.md").write_text(
+    textwrap.dedent(
+        f"""\
+        ---
+        title: "API Reference"
+        ---
+
+        """
+    ),
+    newline="\n",
+)